Day3_Basics_of_GenAI/Task1_Rainbow_Response_HF/requirements.txt
streamlit
langchain-groq
python-dotenv
graphviz
